python3 compatible for cattle counting

+++ Construct_XML_Annotations_cattle_file.py
@@ -8,13 +8,14 @@ through all the subdirectories and restore the files to their original location.
 from cv2 import imread
 from lxml import etree, objectify
 from os import path, mkdir, remove,system
+import os
 from shutil import rmtree, copyfile
 
 from file_helper import read_lines, safe_remove, write_line, read_content
 
 
 def label2xml(list_path, category_id, ANN_DIR='data/Annotations/', IMG_DIR='data/images/', SET_DIR='data/ImageSets/',
-              train=True):
+              train=True, start_img_cnt=0):
     set_file_list = [SET_DIR + 'train.txt', SET_DIR + 'minival.txt', SET_DIR + 'testdev.txt', SET_DIR + 'test.txt']
     TYPEPREFIX = 'train' if train else 'val'
     lines = read_lines(list_path)
@@ -28,29 +29,31 @@ def label2xml(list_path, category_id, ANN_DIR='data/Annotations/', IMG_DIR='data
         infos = line.split()
         image_path = infos[0]
         box = infos[1:]
-        imagename = str(int(image_path.split('.')[0].split('/')[-1]))
+        imagename = int(image_path.split('.')[0].split('/')[-1])
+        dir_num = int(image_path.split('.')[0].split('/')[-2][-5:])
+        target_name = '%05d%04d' % (dir_num, imagename)
 
-        print 'last_path: %s, image_path: %s' % (last_name, image_path)
-        if last_name != imagename:
+        print('last_path: %s, image_path: %s' % (last_name, image_path))
+        if last_name != target_name:
             if train:
-                write_line(set_file_list[0], "%s.jpg %s.xml" % (imagename, imagename))
+                write_line(set_file_list[0], "%s.jpg %s.xml" % (target_name, target_name))
             else:
                 for i in range(1, 4):
-                    write_line(set_file_list[i], "%s.jpg %s.xml" % (imagename, imagename))
+                    write_line(set_file_list[i], "%s.jpg %s.xml" % (target_name, target_name))
             if last_name != '':
                 xml_pretty = etree.tostring(img_annotation, pretty_print=True)
                 with open(ANN_DIR + last_name + ".xml", 'wb') as ann_file:
                     ann_file.write(xml_pretty)
             image_file = imread(image_path)
-            copyfile(image_path, IMG_DIR + '%d.jpg' % int(imagename))
-            if path.exists(ANN_DIR + imagename + ".xml"):
+            copyfile(image_path, IMG_DIR + '%s.jpg' % target_name)
+            if path.exists(ANN_DIR + "%s.xml" % target_name):
                 E = objectify.ElementMaker(annotate=False)
-                img_annotation = objectify.fromstring(read_content(ANN_DIR + imagename + ".xml"))
+                img_annotation = objectify.fromstring(read_content(ANN_DIR + target_name + ".xml"))
             else:
                 E = objectify.ElementMaker(annotate=False)
                 img_annotation = E.annotation(
                     E.folder(TYPEPREFIX),
-                    E.filename(imagename),
+                    E.filename(target_name),
                     E.source(
                         E.database('coco_cattle'),
                     ),
@@ -62,7 +65,7 @@ def label2xml(list_path, category_id, ANN_DIR='data/Annotations/', IMG_DIR='data
                     E.segmented(0)
                 )
 
-            last_name = imagename
+            last_name = target_name
         objectNode = E.object(
             E.name(str(category_id)),
             E.pose("Unspecified"),
@@ -82,18 +85,33 @@ def label2xml(list_path, category_id, ANN_DIR='data/Annotations/', IMG_DIR='data
 def cattle_ssd_prepare(ANN_DIR='data/Annotations/', IMG_DIR='data/images', SET_DIR='data/ImageSets/', force=True):
     set_file_list = [SET_DIR + 'train.txt', SET_DIR + 'minival.txt', SET_DIR + 'testdev.txt', SET_DIR + 'test.txt']
     if force:
-        rmtree(IMG_DIR)
-        rmtree(ANN_DIR)
-        mkdir(IMG_DIR)
-        mkdir(ANN_DIR)
+        if os.path.exists(IMG_DIR):
+            rmtree(IMG_DIR)
+            mkdir(IMG_DIR)
+        else:
+            mkdir(IMG_DIR)
+        if os.path.exists(ANN_DIR):
+            rmtree(ANN_DIR)
+            mkdir(ANN_DIR)
+        else:
+            mkdir(ANN_DIR)
+        if os.path.exists(SET_DIR):
+            rmtree(SET_DIR)
+            mkdir(SET_DIR)
+        else:
+            mkdir(SET_DIR)
+
         for set_file in set_file_list:
             safe_remove(set_file)
     label2xml('data/animal_box.txt', 21, train=True)
     label2xml('data/hard_body.txt', 19, train=True)
+    label2xml('data/hard_body_part.txt', 19, train=True)
     label2xml('data/hard_head.txt', 21, train=True)
+    label2xml('data/hard_head_part.txt', 21, train=True)
     label2xml('data/oid_cattle.txt', 21, train=True)
+    # label2xml('data/oid_cattle_full.txt', 21, train=True)
     label2xml('data/oid_cattle.txt', 21, train=False)
-    system('./data2example.sh')
+    # system('./data2example.sh')
 
 
 if __name__ == '__main__':
